move all Map* attribute tests to their own controller fixture tests
