+++ tests/Functional/Controller/ApiController81.php
@@ -28,16 +28,11 @@ use Nelmio\ApiDocBundle\Tests\Functional\Entity\EntityWithNullableSchemaSet;
 use Nelmio\ApiDocBundle\Tests\Functional\Entity\EntityWithObjectType;
 use Nelmio\ApiDocBundle\Tests\Functional\Entity\EntityWithRef;
 use Nelmio\ApiDocBundle\Tests\Functional\Entity\EntityWithUuid;
-use Nelmio\ApiDocBundle\Tests\Functional\Entity\QueryModel\ArrayQueryModel;
-use Nelmio\ApiDocBundle\Tests\Functional\Entity\QueryModel\FilterQueryModel;
-use Nelmio\ApiDocBundle\Tests\Functional\Entity\QueryModel\PaginationQueryModel;
-use Nelmio\ApiDocBundle\Tests\Functional\Entity\QueryModel\SortQueryModel;
 use Nelmio\ApiDocBundle\Tests\Functional\Entity\RangeInteger;
 use Nelmio\ApiDocBundle\Tests\Functional\Entity\SymfonyConstraints81;
 use Nelmio\ApiDocBundle\Tests\Functional\Entity\SymfonyConstraintsWithValidationGroups;
 use Nelmio\ApiDocBundle\Tests\Functional\Entity\SymfonyDiscriminator81;
 use Nelmio\ApiDocBundle\Tests\Functional\Entity\SymfonyDiscriminatorFileMapping;
-use Nelmio\ApiDocBundle\Tests\Functional\Entity\SymfonyMapQueryString;
 use Nelmio\ApiDocBundle\Tests\Functional\Entity\User;
 use Nelmio\ApiDocBundle\Tests\Functional\EntityExcluded\Symfony7\SerializedNameEntity;
 use Nelmio\ApiDocBundle\Tests\Functional\Form\DummyType;
@@ -48,8 +43,6 @@ use Nelmio\ApiDocBundle\Tests\Functional\Form\FormWithModel;
 use Nelmio\ApiDocBundle\Tests\Functional\Form\FormWithRefType;
 use Nelmio\ApiDocBundle\Tests\Functional\Form\UserType;
 use OpenApi\Attributes as OA;
-use Symfony\Component\HttpKernel\Attribute\MapQueryParameter;
-use Symfony\Component\HttpKernel\Attribute\MapQueryString;
 use Symfony\Component\Routing\Annotation\Route;
 
 class ApiController81
@@ -518,140 +511,5 @@ class ApiController81
     #[OA\Response(response: 200, description: 'Success', content: new Model(type: Dictionary::class))]
     public function dictionary()
     {
-    }
-
-    #[Route('/article_map_query_string')]
-    #[OA\Response(response: '200', description: '')]
-    public function fetchArticleFromMapQueryString(
-        #[MapQueryString] SymfonyMapQueryString $article81Query
-    ) {
-    }
-
-    #[Route('/article_map_query_string_nullable')]
-    #[OA\Response(response: '200', description: '')]
-    public function fetchArticleFromMapQueryStringNullable(
-        #[MapQueryString] ?SymfonyMapQueryString $article81Query
-    ) {
-    }
-
-    #[Route('/article_map_query_string_passes_validation_groups')]
-    #[OA\Response(response: '200', description: '')]
-    public function fetchArticleFromMapQueryStringHandlesValidationGroups(
-        #[MapQueryString(validationGroups: ['test'])] SymfonyConstraintsWithValidationGroups $symfonyConstraintsWithValidationGroups,
-    ) {
-    }
-
-    #[Route('/article_map_query_string_overwrite_parameters')]
-    #[OA\Parameter(
-        name: 'id',
-        in: 'query',
-        schema: new OA\Schema(type: 'string', nullable: true),
-        description: 'Query parameter id description'
-    )]
-    #[OA\Parameter(
-        name: 'name',
-        in: 'query',
-        description: 'Query parameter name description'
-    )]
-    #[OA\Parameter(
-        name: 'nullableName',
-        in: 'query',
-        description: 'Query parameter nullableName description'
-    )]
-    #[OA\Parameter(
-        name: 'articleType81',
-        in: 'query',
-        description: 'Query parameter articleType81 description'
-    )]
-    #[OA\Parameter(
-        name: 'nullableArticleType81',
-        in: 'query',
-        description: 'Query parameter nullableArticleType81 description'
-    )]
-    #[OA\Response(response: '200', description: '')]
-    public function fetchArticleFromMapQueryStringOverwriteParameters(
-        #[MapQueryString] SymfonyMapQueryString $article81Query
-    ) {
-    }
-
-    #[Route('/article_map_query_string_many_parameters')]
-    #[OA\Response(response: '200', description: '')]
-    public function fetchArticleWithManyParameters(
-        #[MapQueryString] FilterQueryModel $filterQuery,
-        #[MapQueryString] PaginationQueryModel $paginationQuery,
-        #[MapQueryString] SortQueryModel $sortQuery,
-        #[MapQueryString] ArrayQueryModel $arrayQuery,
-    ) {
-    }
-
-    #[Route('/article_map_query_string_many_parameters_optional')]
-    #[OA\Response(response: '200', description: '')]
-    public function fetchArticleWithManyOptionalParameters(
-        #[MapQueryString] ?FilterQueryModel $filterQuery,
-        #[MapQueryString] ?PaginationQueryModel $paginationQuery,
-        #[MapQueryString] ?SortQueryModel $sortQuery,
-        #[MapQueryString] ?ArrayQueryModel $arrayQuery,
-    ) {
-    }
-
-    #[Route('/article_map_query_parameter')]
-    #[OA\Response(response: '200', description: '')]
-    public function fetchArticleFromMapQueryParameter(
-        #[MapQueryParameter] int $someInt,
-        #[MapQueryParameter] float $someFloat,
-        #[MapQueryParameter] bool $someBool,
-        #[MapQueryParameter] string $someString,
-        #[MapQueryParameter] array $someArray,
-    ) {
-    }
-
-    #[Route('/article_map_query_parameter_validate_filters')]
-    #[OA\Response(response: '200', description: '')]
-    public function fetchArticleFromMapQueryParameterValidateFilters(
-        #[MapQueryParameter(options: ['min_range' => 2, 'max_range' => 1234])] int $minMaxInt,
-        #[MapQueryParameter(filter: FILTER_VALIDATE_DOMAIN)] string $domain,
-        #[MapQueryParameter(filter: FILTER_VALIDATE_EMAIL)] string $email,
-        #[MapQueryParameter(filter: FILTER_VALIDATE_IP)] string $ip,
-        #[MapQueryParameter(filter: FILTER_VALIDATE_IP, flags: FILTER_FLAG_IPV4)] string $ipv4,
-        #[MapQueryParameter(filter: FILTER_VALIDATE_IP, flags: FILTER_FLAG_IPV6)] string $ipv6,
-        #[MapQueryParameter(filter: FILTER_VALIDATE_MAC)] string $macAddress,
-        #[MapQueryParameter(filter: FILTER_VALIDATE_REGEXP, options: ['regexp' => '/^test/'])] string $regexp,
-        #[MapQueryParameter(filter: FILTER_VALIDATE_URL)] string $url,
-    ) {
-    }
-
-    #[Route('/article_map_query_parameter_nullable')]
-    #[OA\Response(response: '200', description: '')]
-    public function fetchArticleFromMapQueryParameterNullable(
-        #[MapQueryParameter] ?int $id,
-    ) {
-    }
-
-    #[Route('/article_map_query_parameter_default')]
-    #[OA\Response(response: '200', description: '')]
-    public function fetchArticleFromMapQueryParameterDefault(
-        #[MapQueryParameter] int $id = 123,
-    ) {
-    }
-
-    #[Route('/article_map_query_parameter_overwrite_parameters')]
-    #[OA\Parameter(
-        name: 'id',
-        in: 'query',
-        description: 'Query parameter id description',
-        example: 123,
-    )]
-    #[OA\Parameter(
-        name: 'changedType',
-        in: 'query',
-        schema: new OA\Schema(type: 'int', nullable: false),
-        description: 'Incorrectly described query parameter',
-        example: 123,
-    )]
-    #[OA\Response(response: '200', description: '')]
-    public function fetchArticleFromMapQueryParameterOverwriteParameters(
-        #[MapQueryParameter] ?int $id,
-        #[MapQueryParameter] ?string $changedType,
-    ) {
     }
 }
